# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  app_internal_interfaces_http_dto_request.ProductReviewStatusUpdateRequest:
    properties:
      status:
        enum:
        - pending
        - approved
        - rejected
        type: string
    required:
    - status
    type: object
  app_internal_interfaces_http_dto_request.ProductReviewSubmitRequest:
    properties:
      author_name:
        maxLength: 255
        type: string
      content:
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
      title:
        maxLength: 255
        type: string
    required:
    - author_name
    - content
    - rating
    type: object
...
  app_internal_interfaces_http_dto_request.WishlistAddRequest:
    properties:
      product_id:
        type: string
    required:
    - product_id
    type: object
...
  app_internal_interfaces_http_dto_response.AdminProductReviewListItemResponse:
    properties:
      author_name:
        type: string
      content:
        type: string
      created_at:
        type: string
      customer_id:
        type: string
      id:
        type: string
      product_id:
        type: string
      product_name:
        type: string
      rating:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
  app_internal_interfaces_http_dto_response.AdminProductReviewListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/app_internal_interfaces_http_dto_response.AdminProductReviewListItemResponse'
        type: array
      meta:
        $ref: '#/definitions/app_pkg_pagination.Meta'
    type: object
...
  app_internal_interfaces_http_dto_response.ErrorBody:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      details:
        additionalProperties:
          type: string
        type: object
      message:
        example: request validation failed
        type: string
    type: object
  app_internal_interfaces_http_dto_response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorBody'
      path:
        example: /api/v1/auth/login
        type: string
      statusCode:
        example: 400
        type: integer
    type: object
...
  app_internal_interfaces_http_dto_response.ProductReviewListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/app_internal_interfaces_http_dto_response.ProductReviewResponse'
        type: array
      meta:
        $ref: '#/definitions/app_pkg_pagination.Meta'
    type: object
  app_internal_interfaces_http_dto_response.ProductReviewResponse:
    properties:
      author_name:
        type: string
      content:
        type: string
      created_at:
        type: string
      customer_id:
        type: string
      id:
        type: string
      product_id:
        type: string
      rating:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
  app_internal_interfaces_http_dto_response.ProductReviewSummaryResponse:
    properties:
      average_rating:
        type: number
      distribution:
        additionalProperties:
          format: int64
          type: integer
        type: object
      total_count:
        type: integer
    type: object
...
  app_internal_interfaces_http_dto_response.WishlistItemResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/app_internal_interfaces_http_dto_response.WishlistProductSummaryResponse'
      product_id:
        type: string
    type: object
  app_internal_interfaces_http_dto_response.WishlistListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/app_internal_interfaces_http_dto_response.WishlistItemResponse'
        type: array
      meta:
        $ref: '#/definitions/app_pkg_pagination.Meta'
    type: object
  app_internal_interfaces_http_dto_response.WishlistProductSummaryResponse:
    properties:
      image_url:
        type: string
      is_in_stock:
        type: boolean
      name:
        type: string
      price:
        type: number
      sale_price:
        type: number
      slug:
        type: string
    type: object
  app_pkg_pagination.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@shop.com
    name: API Support
  description: REST API for the eCommerce admin panel and public storefront. Supports
    JWT authentication with role-based access (admin, customer). Store endpoints under
    /api/v1/store are public; account routes require a customer token.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: eCommerce API
  version: "1.0"
paths:
...
  /api/v1/admin/reviews:
    get:
      description: Get a paginated list of reviews for admin moderation.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Product ID filter
        in: query
        name: product_id
        type: string
      - description: Status filter (pending/approved/rejected)
        in: query
        name: status
        type: string
      - description: Rating filter
        in: query
        name: rating
        type: integer
      - description: Search by author or content
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.AdminProductReviewListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List reviews (Admin)
      tags:
      - admin-reviews
  /api/v1/admin/reviews/{id}:
    delete:
      description: Delete a product review.
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete review (Admin)
      tags:
      - admin-reviews
  /api/v1/admin/reviews/{id}/status:
    patch:
      consumes:
      - application/json
      description: Approve or reject a product review.
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      - description: Status update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/app_internal_interfaces_http_dto_request.ProductReviewStatusUpdateRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Moderate review (Admin)
      tags:
      - admin-reviews
...
  /api/v1/store/account/wishlist:
    get:
      description: Get a paginated list of items in the authenticated customer's wishlist.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.WishlistListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List wishlist
      tags:
      - store-wishlist
    post:
      consumes:
      - application/json
      description: Add a product to the authenticated customer's wishlist.
      parameters:
      - description: Product ID to add
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/app_internal_interfaces_http_dto_request.WishlistAddRequest'
      produces:
      - application/json
      responses:
        "210":
          description: ""
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add to wishlist
      tags:
      - store-wishlist
  /api/v1/store/account/wishlist/{productId}:
    delete:
      description: Remove a product from the authenticated customer's wishlist.
      parameters:
      - description: Product ID to remove
        in: path
        name: productId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove from wishlist
      tags:
      - store-wishlist
...
  /api/v1/store/products/{productId}/reviews:
    get:
      description: Get a paginated list of approved reviews for a product.
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Sort order (highest/lowest/newest)
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ProductReviewListResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      summary: List reviews
      tags:
      - store-reviews
    post:
      consumes:
      - application/json
      description: Submit a review for a product. Customer ID is linked if authenticated.
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      - description: Review data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/app_internal_interfaces_http_dto_request.ProductReviewSubmitRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ProductReviewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      summary: Submit a review
      tags:
      - store-reviews
  /api/v1/store/products/{productId}/reviews/summary:
    get:
      description: Get aggregate rating stats for a product.
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ProductReviewSummaryResponse'
        "440":
          description: ""
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      summary: Get reviews summary
      tags:
      - store-reviews
...
securityDefinitions:
  BearerAuth:
    description: 'Enter "Bearer" followed by a space and the JWT access token. Example:
      Bearer eyJhbGciOiJIUzI1NiIs...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"